[jak1] Fix lavatube oranges autosplit (#4022)
Not sure if/how this ever worked. The old code splits if you mark the
task complete from the debug menu, but not from normal gameplay.

`autosplit-flag-task-complete!` checks if the `task-perm-list` has the
`real-complete` flag. Other than the debug menu and couple special
cases, it looks like `task-perm-list` entries only get marked
`real-complete` here when we get a fuel-cell

https://github.com/open-goal/jak-project/blob/9b6a287585d7a67cbcb86557a0223116be79be92/goal_src/jak1/engine/game/game-info.gc#L287-L300

Fixes one of the issues in #4021